Extracted mouse handlers from the TableSelection plugin to the new TableMouse plugin.
