+++ packages/ckeditor5-table/src/tablemouse/mouseeventsobserver.js
